AGENCIAS/backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic==2.9.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
requests==2.32.3
httpx==0.27.0
PyJWT>=2.9